refactor: BoardRepository 인터페이스 생성 및 분리, Jpa 구현
